+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Changes:
 
+- `createType<T = Codec>('...': K)` now returns `Codec` when `K` is not in the InterfaceRegistry, with `T` override
 - Handle `BitSequence` in `PortableRegistry`
 - Move all static test support files into `@polkadot/types-support`
 - Align static Metadata with latest Substrate, Polkadot & Kusama